BB: Move mimikatz test to depth 2a
Depth 2 is required to steal and use the stolen credentials. If the depth is only 1, the island will not be able to exploit mimikatz-15 with stolen credentials, because mimikatz-14
